+++ core/evol/method/mutate/DoerrMutation.cpp
@@ -6,8 +6,8 @@ namespace ea::method {
 
 DoerrMutation::DoerrMutation(DoerrMutationConfig const& config) : beta_(config.beta()) {}
 
-double DoerrMutation::_get_alpha(size_t size) const {
-  size_t bound = size / 2 + 1;
+double DoerrMutation::_get_alpha(uint32_t size) const {
+  uint32_t bound = size / 2 + 1;
   if (bound < 3) {
     return 1.;
   }
@@ -16,11 +16,11 @@ double DoerrMutation::_get_alpha(size_t size) const {
   double p = core::random::sample<double>(0., 1.);
   double c = 0.;
 
-  for (size_t i = 1; i < bound; ++i) {
+  for (uint32_t i = 1; i < bound; ++i) {
     c += 1. / std::pow(i, beta_);
   }
 
-  for (size_t k = 1; k < bound; ++k) {
+  for (uint32_t k = 1; k < bound; ++k) {
     ll = rr;
     rr += (1. / (c * std::pow(k, beta_)));
     if (ll <= p && p < rr) {
@@ -32,13 +32,13 @@ double DoerrMutation::_get_alpha(size_t size) const {
 }
 
 void DoerrMutation::apply(instance::Instance& instance) {
-  size_t num_vars = instance.num_vars();
+  uint32_t num_vars = instance.num_vars();
   if (alpha_ == -1.) {
     alpha_ = _get_alpha(num_vars);
   }
   double p = alpha_ / (double) num_vars;
 
-  for (size_t i = 0; i < num_vars; ++i) {
+  for (uint32_t i = 0; i < num_vars; ++i) {
     if (core::random::flip_coin(p)) {
       instance.get_vars().flip(i);
     }
