+++ core/evol/method/mutate/UniformMutation.cpp
@@ -7,9 +7,9 @@ namespace ea::method {
 UniformMutation::UniformMutation(UniformMutationConfig const& config) : scale_(config.scale()) {}
 
 void UniformMutation::apply(instance::Instance& instance) {
-  size_t num_vars = instance.num_vars();
+  uint32_t num_vars = instance.num_vars();
   double p = scale_ / (double) num_vars;
-  for (size_t i = 0; i < num_vars; ++i) {
+  for (uint32_t i = 0; i < num_vars; ++i) {
     if (core::random::flip_coin(p)) {
       instance.get_vars().flip(i);
     }
